fixed index-files, and moved formField to containers
