+++ source/components/organisms/Step/Step.js
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components/native';
 import AuthContext from 'app/store/AuthContext';
 import { Text } from 'react-native';
-import { AuthLoading, FormField } from 'app/components/molecules';
+import { AuthLoading } from 'app/components/molecules';
+import FormField from 'app/containers/FormField';
 import BackNavigation from '../../molecules/BackNavigation/BackNavigation';
 import Banner from './StepBanner/StepBanner';
 import StepFooter from './StepFooter/StepFooter';
