+++ source/components/organisms/SummaryList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SummaryList';
