--- source/components/organisms/index.js
@@ -1,1 +1,0 @@
-export { default as Step } from './Step';
