+++ source/components/organisms/index.ts
@@ -1,0 +1,3 @@
+export { default as Step } from './Step';
+export { default as SummaryList } from './SummaryList';
+export { default as YoutubePopup } from './YoutubePopup';
